docs!: require Ruby users to go through OpenAPI, instead of through the gem (#956)
PSA to Ruby users: if you currently rely on the `baml` gem, please reach
out to us either in the Discord or on GitHub so we can migrate you to
the OpenAPI-generated Ruby client.

Over the next few weeks, we plan to provide white-glove support to
migrate all Ruby users from the `baml` gem to the OpenAPI-generated
client and move the `baml` gem into maintenance mode. This will mean
that:

- we will remove `baml`-gem-specific documentation from
https://docs.boundaryml.com/
- we will only publish security fixes for the `baml` gem
- we will ask that you have a plan to migrate to the OpenAPI-generated
Ruby client, before providing any support for

## Why are we doing this?

Bluntly speaking: we don’t think we can provide users of the `baml` gem
with a good experience. It’s in a pretty sorry state today:

- it doesn't work on Windows
- it doesn't work out of the box on macOS (some issue with `darwin` vs
`darwin23`?)
- it still doesn't support tracing
- it's incompatible with async (the underlying FFI libraries do not make
this easy, and we don’t have the engineering resources to do this well)

We're also not aware of anyone _using_ the `baml` gem today, we suspect
in large part because of how buggy it is.

Due to the combination of all of these issues, we will be dropping the
`baml` gem in favor of the OpenAPI-generated Ruby client.