fix(vscode): allow json files in baml_src (#1513)
BAML, a long time ago, used to allow defining tests in JSON files, so we
included them in baml_src.

Now that tests are defined exclusively in baml files, the extension
implodes if there are .json files in baml_src.

This is in direct conflict with the new playground proxy setting toggle,
which creates and updates `.vscode/settings.json` in baml_src if you
have that folder open.

---------

Co-authored-by: ellipsis-dev[bot] <65095814+ellipsis-dev[bot]@users.noreply.github.com>