feat: add explicit handling for newlines in env vars (#1775)
This will not allow copy-pasting values with newlines into a given env
var input, but you can do that using the "import a .env file" field. I
could use textarea but don't yet want to compromise on accessibility /
hand-rolling input type=password with that, and am not sure what npm
packages exist for that.