Fix Tracing bug and LSP Crash (#2030)
Fix an issue where BAML_LOG wouldn't be printed to the console in async
contexts in python due to the context stack not being propagated
correctly.

Log an error when we run into issues with the tracing spans

Fix an issue where LSP would crash when sending diagnostics for a baml
file not in a baml_src.