[BREAKING] Reserve BAML keywords (#2473)
Reserve a set of BAML language keywords, so they are not admissible in
user code as parameter names, function names or local variable names.

This is a breaking change for any users who have previously named their
parameter in a way that overlaps with reserved keywords, such as `image`
(which appears as a parameter name in our own test suite).