Bedrock modular api (#2526)
Modular API didn't support Bedrock because a Bedrock relies on the
aws-sdk for constructing and signing requests - HTTP requests aren't
naturally exposed to the user with that SDK.

To get around this, we take an intermediate object and serialize that to
JSON to produce something resembling an HTTP request that users can
modify.

To send the request after modification, it must be signed again. We have
new docs describing this process.