fix(playground): make vertex work in the vscode playground (#2525)
The vertex provider currently only works in the playground if the user
has run both `gcloud auth application-default login` and `gcloud auth
login` - but only the former should actually be required to resolve
sufficient credentials for the application to work.

In the case where the user has not run `gcloud auth login`, then in the
extension host `GoogleAuth.getProjectId()` will fail because there is no
source to load a project ID from (it is possible to load a project ID
via ADC if the ADC is sourced from _service account_ creds, but not when
using a user's default creds).

This updates the comms path by which the playground retrieves GCP creds
to allow project ID loading to fall back to resolving from the oauth
token (I think that's what's happening, because it's still getting
resolved, just seems to be at a later stage).

Error message still tells the user to run `gcloud auth
application-default login`, but we've also added more specficity to make
disambiguating error message sites easier.