Implement streaming for emit values (#2595)
This works (although it uses the toy emitter from #2593, real dependency
tracking in the VM is WIP), meaning that when values are emitted,
streams are emitted too, and when the stream is from an assignment of an
LLM function return value to a variable, the chunks are streamed.

I will take some time to see if it can be simplified (the implementation
is more complicated than I'd like).