Update Fern CLI to configure settings from docs.yml (#2606)
These settings were previously controlled in an internal configuration.
Now, we have transferred these settings to be controllable from the
`docs.yml`. This PR adds the settings currently enabled in the internal
configuration for Boundary to the `docs.yml` file.

We will be deprecating the internal configuration in the coming weeks.
Please merge to avoid any disruption in these settings.

---------

Signed-off-by: Fern Support <126544928+fern-support@users.noreply.github.com>