Fix timeout exceptions for streaming LLM calls (#2699)
Prior to this PR, the `request_timeout_ms` was implemented in a way that
quietly cancelled the LLM response.

This PR checks, after the response finishes, whether or not the LLM
provider's stop signal was received by the runtime, and when not
received, returns a response failures, which gets turned into the
appropriate exception downstream.