Reorganize the error formatting code (#2751)
Here's an experiment in organizing the error formatting code.

We're optimizing for:
 - User-facing messages are written in a single file
 - Compiler phases get to define their own non-error types
 - Compiler phases get to _have_ phase-specific errors

The part to focus on for maintaining the error messages is here:

https://github.com/BoundaryML/baml/compare/greg/centralize-tests?expand=1#diff-ad8930bb791d17695f5083d16f7c9d3542646eda5373033e6a447040040d0ce0R11

The organization now is:

 - No `Diagnostics` trait
- A new `compiler_errors` module that collects every error type for
every compiler phase
- A single `error_format.rs` file that knows how to construct an error
message for each error type
- Error rendering goes straight to an `ariande::ReportBuilder` - this is
because different error messages will need different report shapes. We
don't want to drive it all the way to a single message.

Some unusual consequences from following the constraints:
- The error types aren't defined in the crate that defines the compiler
phase
- Sometimes error types need a type parameter, to resolve cyclic Rust
dependencies. For example, `baml_thir` wants its errors to contain
types. But `baml_thir` depends on `baml_diagnostics` (where the errors
are defined), so `baml_diagnostics` can not depend on `baml_thir`...
therefore `baml_diagnostics` can't directly mention `Ty` (the definition
of types defined in `baml_thir`.